ParallelKMeans.jl

ParallelKMeans.divider

ParallelKMeans.inner_pairwise!

ParallelKMeans.kmeans

ParallelKMeans.pairwise!

ParallelKMeans.smart_init

ParallelKMeans.sum_of_squares

ParallelKMeans.kmeans — Method

Kmeans(design_matrix, k; k_init="k-means++", max_iters=300, tol=1e-4, verbose=true)

This main function employs the K-means algorithm to cluster all examples in the training data (design_matrix) into k groups using either the k-means++ or random initialisation technique for selecting the initial centroids.

At the end of the number of iterations specified (max_iters), convergence is achieved if difference between the current and last cost objective is less than the tolerance level (tol). An error is thrown if convergence fails.

Details of operations can be either printed or not by setting verbose accordingly.

A tuple representing labels, centroids, and sum_squares respectively is returned.

ParallelKMeans.divider — Method

divider(n, k)

Utility function, splits 1:n sequence to k chunks of approximately same size.

ParallelKMeans.inner_pairwise! — Method

inner_pairwise!(target, x, y, r)

Utility function for calculation of pairwise!(target, x, y, mode) function. UnitRange argument r select subarray of original design matrix x that is going to be processed.

ParallelKMeans.pairwise! — Method

pairwise!(target, x, y, mode)

Let X and Y respectively have m and n columns. Then the pairwise! function computes distances between each pair of columns in X and Y and store result in target array. Argument mode defines calculation mode, currently following modes supported

SingleThread()

MultiThread()

ParallelKMeans.smart_init — Method

smart_init(X, k; init="k-means++")

This function handles the random initialisation of the centroids from the design matrix (X) and desired groups (k) that a user supplies.

k-means++ algorithm is used by default with the normal random selection of centroids from X used if any other string is attempted.

A tuple representing the centroids, number of rows, & columns respecitively is returned.

ParallelKMeans.sum_of_squares — Method

sum_of_squares(x, labels, centre, k)

This function computes the total sum of squares based on the assigned (labels) design matrix(x), centroids (centre), and the number of desired groups (k).

A Float type representing the computed metric is returned.